stars: Explicitly cast RAND_MAX to float
Some platforms warn about a loss of precision without an explicit cast.
We don't care too much about precision, so we can explicitly cast to
avoid the warning.

    stars.c:59:38: error: implicit conversion from 'int' to 'float' changes value from 2147483647 to 2147483648 [-Werror,-Wimplicit-const-int-float-conversion]
        const float t = ((float)rand() / RAND_MAX) * (2.0f * M_PI);
                                       ~ ^~~~~~~~
    /opt/homebrew/Cellar/emscripten/2.0.15/libexec/cache/sysroot/include/stdlib.h:82:19: note: expanded from macro 'RAND_MAX'
    #define RAND_MAX (0x7fffffff)
                      ^~~~~~~~~~
    stars.c:60:48: error: implicit conversion from 'int' to 'float' changes value from 2147483647 to 2147483648 [-Werror,-Wimplicit-const-int-float-conversion]
        const float r = r_range * ((float)rand() / RAND_MAX);
                                                 ~ ^~~~~~~~
    /opt/homebrew/Cellar/emscripten/2.0.15/libexec/cache/sysroot/include/stdlib.h:82:19: note: expanded from macro 'RAND_MAX'
    #define RAND_MAX (0x7fffffff)
                      ^~~~~~~~~~
    stars.c:67:52: error: implicit conversion from 'int' to 'float' changes value from 2147483647 to 2147483648 [-Werror,-Wimplicit-const-int-float-conversion]
        star->initial_color = (1.0f - ((float)rand() / RAND_MAX) * 0.75f) * 255.0f;
                                                     ~ ^~~~~~~~
    /opt/homebrew/Cellar/emscripten/2.0.15/libexec/cache/sysroot/include/stdlib.h:82:19: note: expanded from macro 'RAND_MAX'
    #define RAND_MAX (0x7fffffff)
                      ^~~~~~~~~~
    3 errors generated.

+++ stars/stars.c
@@ -56,15 +56,15 @@ struct star stars[NUM_STARS];
 static void init_star(struct star *star)
 {
     /* Make a disc of stars. */
-    const float t = ((float)rand() / RAND_MAX) * (2.0f * M_PI);
-    const float r = r_range * ((float)rand() / RAND_MAX);
+    const float t = ((float)rand() / (float)RAND_MAX) * (2.0f * M_PI);
+    const float r = r_range * ((float)rand() / (float)RAND_MAX);
     star->x = r * cos(t);                 /* [-r_range, r_range] */
     star->y = r * sin(t);                 /* [-r_range, r_range] */
     star->z = z_start;
 
 
     /* [0, 255] */
-    star->initial_color = (1.0f - ((float)rand() / RAND_MAX) * 0.75f) * 255.0f;
+    star->initial_color = (1.0f - ((float)rand() / (float)RAND_MAX) * 0.75f) * 255.0f;
     star->initial_color = star->initial_color > 255 ? 255
                                                     : star->initial_color;
 }
